daemon: Don't forward NP req over wg if no skb_change_head
This commit disables NodePort BPF requests forwarding to remote nodes
over the Wireguard tunnel if the "skb_change_head()" BPF helper is not
available (the case for < 5.8 kernels).

Without the helper it is not possible to create the headroom for L2 hdr.
Unfortunately, "skb_adjust_room(skb, 14, BPF_ADJ_ROOM_MAC, 0)" cannot be
used either, as it does not set "skb->mac_header" which makes the packet
to be dropped by [1] when redirecting to another device.

[1]: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/net/core/filter.c?h=v5.8#n2118

Signed-off-by: Martynas Pumputis <m@lambda.lt>