spiffe: automatically recover from errors
This commit improves error handling by:

- Tracking the expiration time of the spiffe-based labels. Spiffe-based
labels are removed from the endpoint if they are not renewed after they
expire. This is useful to handle situation where the spire control plane
is down and cilium-agent doesn't know the status of those entries.

- From Mauricio's commit 1b29045cd4bc87f439d92ef3a834587d69eb0fd5,
change some points regarding the single watch approach of the Spire API.